Bubble up CGroup::GetCpuLimit to callers of GetCurrentProcessCpuCount that can take advante of it
In this case, it's only the ThreadPool. It's used to make sure we optimize the number of worker threads allocated based on the CPU budget.

+++ src/pal/inc/pal.h
@@ -2454,7 +2454,7 @@ PAL_GetPhysicalMemoryUsed(size_t* val);
 PALIMPORT
 BOOL
 PALAPI
-PAL_GetCpuLimit(UINT* val);
+PAL_GetCpuLimit(double* val);
 
 PALIMPORT
 size_t
